+++ plugins/11_TheWarWithin/localization/zhCN.lua
@@ -677,3 +677,5 @@ L['ancient_coffer_note'] = '收集 {item:233794}（{dot:Red}）和 {item:233799}
 
 L['sthaarbs_note'] = '使用 {object:相位导管} 附近的传送器前往 {npc:234845} 处。'
 L['heka_tamos_note'] = '从附近的 {object:望露莲}、{object:幽灵提灯}、{object:土生肉质植物} 和 {object:捕风花} 处各收集一层增益效果，然后与 {object:元素团结火盆} 互动召唤稀有。'
+
+L['translocated_gorger_note'] = nil